% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{AMS}
\alias{AMS}
\title{Alpha shape parameter given mean and variance of a Standard Beta PDD.}
\usage{
AMS(M, VAR)
}
\arguments{
\item{M}{The mean of the target Standard Beta probability density distribution.}

\item{VAR}{The variance of the target Standard Beta probability density distribution.}
}
\description{
Calculates the Alpha value required to produce a Standard Beta probability density distribution with defined mean and variance.
}
