% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{AUC}
\alias{AUC}
\title{Area Under the ROC Curve.}
\usage{
AUC(FPR, TPR)
}
\arguments{
\item{FPR}{Vector of False-Positive Rates.}

\item{TPR}{Vector of True-Posiitive Rates.}
}
\value{
A value representing the area under the ROC curve.
}
\description{
Given a vector of false-positive rates and a vector of true-positive rates, calculate the area under the Receiver Operator Characteristic (ROC) curve.
}
\note{
Script originally retrieved and modified from https://blog.revolutionanalytics.com/2016/11/calculating-auc.html.
}
