% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{Beta.4p.fit}
\alias{Beta.4p.fit}
\title{Method of Moment Estimates of Shape- and Location Parameters of the Four-Parameter Beta Distribution.}
\usage{
Beta.4p.fit(scores)
}
\arguments{
\item{scores}{A vector of values to which the four-parameter beta distribution is to be fitted.}
}
\value{
A list of parameter-values required to produce a four-parameter beta distribution with the same first four moments as the observed distribution.
}
\description{
An implementation of the method of moments estimation of four-parameter beta distribution parameters presented by Hanson (1991). Given a string of values, calculates the shape- and location parameters required to produce a four-parameter beta distribution with the same mean, variance, skewness and kurtosis (i.e., the first four moments) as the observed-score distribution.
}
\references{
Hanson, Bradley A. (1991). Method of Moments Estimates for the Four-Parameter Beta Compound Binomial Model and the Calculation of Classification Consistency Indexes.American College Testing Research Report Series.

Lord, Frederic M. (1965). A Strong True-Score Theory, With Applications. Psychometrika, 30(3).
}
