% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{Beta.gfx.poly}
\alias{Beta.gfx.poly}
\title{Coordinate generation for marking an area under the curve for the Standard Beta probability density distribution.}
\usage{
Beta.gfx.poly(FROM, TO, BY, Alpha, Beta)
}
\arguments{
\item{FROM}{The point of the x-axis from where to start producing y-density values.}

\item{TO}{The point of the x-axis to where y-density values are to be produced.}

\item{BY}{The resolution (or specing) at which to produce y-density values.}

\item{Alpha}{The Alpha shape-parameter value for the Standard Beta probability density distribution.}

\item{Beta}{The Beta shape-parameter fort he Standard Beta probability density distribution.}
}
\value{
A two-column matrix with density-values of y to plot against corresponding location values of x.
}
\description{
Plotting tool, producing a two-column matrix with values of \code{y} corresponding to locations on \code{x}. Useful for shading areas under the curve when tracing the line for the Standard Beta probability density function.
}
