% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{dBeta.4P}
\alias{dBeta.4P}
\title{Probability Density under the Four-Parameter Beta PDD.}
\usage{
dBeta.4P(x, l, u, alpha, beta)
}
\arguments{
\item{x}{Value of X.}

\item{l}{The first (lower) location parameter.}

\item{u}{The second (upper) location paraeter.}

\item{alpha}{The first shape parameter.}

\item{beta}{The second shape parameter.}
}
\value{
The value for the probability density at specified values of X.
}
\description{
Gives the density at desired values of X under the Four-Parameter Beta PDD.
}
\examples{
# Assume some variable follows a four-parameter beta distribution with
# location parameters \\code{l = 0.25} and \\code{u = .75}, and shape
# parameters \\code{alpha = 5} and \\code{beta = 3}. To compute the
# probability density at a specific point of the distribution (e.g., .5)
# using \\code{dBeta.4P()}:
dBeta.4P(x = .5, l = .25, u = .75, alpha = 5, beta = 3)
}
