% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{DDG}
\alias{DDG}
\title{Dichotomous Data Generator.}
\usage{
DDG(ipt, theta)
}
\arguments{
\item{theta}{A vector of values representing respondents standing on the latent variable, where vector length is equal to the number of respondents responding to the items.}

\item{iptable}{A table where the number of rows equals the number of items, and the columns represent the items' difficulty ([, 1]), discrimination ([, 2]), and guessing ([, 3]) parameters, respectively.}
}
\value{
A nRespondent x nItem data-frame of dichotomous responses, probabilistically generated in accordance with the respondents' probability of correct response under some previously specified IRT model.
}
\description{
generated a data-frame where ncol = number of items and nrow = number of respondents. Responses take the form of dichotomous values (0/1), where the values of each individual observation is probabilistically generated based on the respondents probability of responding "1" to the specific item with the properties "difficulty" ("b"), "discrimination" ("a"), and "guessing" ("c").
}
