% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{ETL}
\alias{ETL}
\title{Livingston and Lewis' "Effective Test Length"}
\usage{
ETL(mean, variance, l = 0, u = 1, reliability)
}
\arguments{
\item{mean}{The mean of the observed-score distribution.}

\item{variance}{The variance of the observed-score distribution.}

\item{l}{The lower-bound of the observed-score distribution.}

\item{u}{The upper-bound of the observed-score distribution.}

\item{reliability}{The reliability of the observed scores (correlation with true-score distribution).}
}
\value{
An estimate of the effective length of a test, given the stability of the observations it produces.
}
\description{
According to Livingston and Lewis (1995), "The effective test length corresponding to a test score is the number of discrete, dichotomously scored, locally independent, equally difficult items required to produce a total score of the same reliability."
}
\references{
Livingston, Samuel A. and Lewis, Charles. (1995). Estimating the Consistency and Accuracy of Classifications Based on Test Scores. Journal of Educational Measurement, 32(2).
}
