% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{HHMLE}
\alias{HHMLE}
\title{Heavy-Handed Maximum Likelihood Estimation of Theta.}
\usage{
HHMLE(rp, ip, ThetaRange = c(-6, 6), resolution = 0.001)
}
\arguments{
\item{rp}{A vector representing a specific response-pattern.}

\item{ip}{A table of item properties (columns representing difficulty, discrimination, and guessing parameters, respectively).}

\item{ThetaRange}{The range across which Theta values are to be evaluated with respect to their probability of producing the response pattern.}

\item{resoluation}{The resolution at which Theta is to be evaluated (i.e., desired decimal-level of precision).}
}
\description{
Estimates Theta given supplied response-pattern and item properties by evaluating the log-likelihood of Theta across the Theta spectrum and choosing the value of Theta corresponding to the largest value.
}
