% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{IPG}
\alias{IPG}
\title{Item Properties Generator.}
\usage{
IPG(nitem = 10, b.range = c(-3, 3), rndm.b = FALSE, a.range = c(1,
  1), rand.a = TRUE, c.range = c(0, 0), rand.c = TRUE)
}
\arguments{
\item{nitem}{The number of items for which item parameters are to be generated.}

\item{b.range}{The range to be covered by the difficulty parameters.}

\item{rndm.b}{Logical. Whether item difficulty parameters are to be uniformly randomly distributed in the specified range, or whether items are to cover the range with equal spacing between each item. Defaults to FALSE (equal spacing).}

\item{a.range}{The range to be covered by the discrimination parameters.}

\item{rand.a}{Logical. Whether item discrimination parameters are to be uniformly randomly distributed in the specified range, or whether items are to cover the range with equal spacing. Defaults to TRUE (uniform random distribution).}

\item{c.range}{The range to be covered by the discrimination parameters.}

\item{rand.c}{Logical. Whether item guessing parameters are to be uniformly randomly distributed in the specified range, or whether items are to cover the range with equal spacing. Defaults to FALSE (uniform random distribution).}
}
\value{
A matrix of item properties, where [, 1] represents difficulty, [, 2] discrimination, and [, 3] guessing.
}
\description{
Generates a matrix of dichotomous-item IRT-parameters (discrimination, difficulty, and guessing), where the first column represents the difficulty parameter, the second the discrimination parameter, and the thirt the guessing parameter for the item.
}
