% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{LL.CA}
\alias{LL.CA}
\title{An implementation of the Livingston and Lewis (1995) Method for estimating Classification Accuracy based on Test Scores.}
\usage{
LL.CA(x = NULL, reliability, cut, min = 0, max = 1,
  error.model = "binomial", truecut = NULL, grainsize = 0.001)
}
\arguments{
\item{x}{A vector of observed scores for which a beta-distribution is to be fitted.}

\item{reliability}{The observed-score correlation with the true-score.}

\item{cut}{The cutoff value for classifying observations into pass or fail categories.}

\item{min}{The minimum value possible to attain on the test. Default is 0.}

\item{max}{The maximum value possible to attain on the test. Default is 1.}

\item{error.model}{The probability distribution to be used for producing the sampling distributions at different points of the true-score scale. Options are \code{beta} and \code{binomial}. The binomial distribution is discrete, and is the distribution used originally by Livingston and Lewis. Use of the binomial distribution involves a rounding of the effective test length to the nearest integer value. The Beta distribution is continuous, and does not involve rounding of the effective test length..}

\item{truecut}{Optional specification of a "true" cutoff. Useful for producing ROC curve values.}

\item{grainsize}{The size of the steps for which probabilities along the score distribution are to be calculated. Default is .001 (1001 points).}
}
\value{
A confusion matrix estimating the proportion of true/false pass/fail categorizations for a test, given a specific distribution of observed scores.
}
\description{
An implementation of what has been come to be known as the "Livingston and Lewis approach" to classification accuracy, which assumes that observed-scores, true-scores, and errors of measurement follow the four-parameter beta distribution. Under this assumption, the expected classification consistency and accuracy of tests can be estimated from observed outcomes and estimated test reliability.
}
\references{
Livingston, Samuel A. and Lewis, Charles. (1995). Estimating the Consistency and Accuracy of Classifications Based on Test Scores. Journal of Educational Measurement, 32(2).
}
