% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{LNLRV}
\alias{LNLRV}
\title{The Natural-Log of the Likelihood of a Response Vector given a value of Theta.}
\usage{
LNLRV(x, iptable, theta)
}
\arguments{
\item{x}{A vector of dichotomous values (i.e., 0/1) with length equal to the number of items.}

\item{iptable}{A table of item-parameters, where the first collumn represents the difficulty parameters, the second the discrimination parameters, and the thirt the guessing parameters.}

\item{theta}{A specific value of Theta ("Ability") for which the probability of the supplied response vector is to be calculated.}
}
\value{
The log of the likelihood of producing the supplied response vector at a given value of Theta.
}
\description{
Calculates the natoral logarithm of the likelihood of an individual with ability-value Theta attaining a specific response vector given a set of item-parameters, assuming local independence.
}
