% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{MLA}
\alias{MLA}
\title{Most likely true alpha value given observed outcome.}
\usage{
MLA(a, b, x = NULL, N = NULL)
}
\arguments{
\item{a}{Observed alpha value for fitted Standard Beta PDD.}

\item{b}{Observed beta value for fitted Standard Beta PDD.}
}
\value{
The Alpha shape-parameter value for the Standard Beta probability density distribution where the observed mean is the expected mode.
}
\description{
Given a fitted Standard Distribution, return the Alpha value where the observed mean becomes the mode.
}
