% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{MLB}
\alias{MLB}
\title{Most likely beta value given observed outcome.}
\usage{
MLB(a, b)
}
\arguments{
\item{a}{Observed alpha value for fitted Standard Beta PDD.}

\item{b}{Observed beta value for fitted Standard Beta PDD.}
}
\value{
The Beta shape-parameter value for the Standard Beta probability density distribution where the observed mean is the expected mode.
}
\description{
Given a fitted Standard Beta Distribution, return the Beta value where the observed mean becomes the mode.
}
