% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{NMLE}
\alias{NMLE}
\title{Newton's Method of Maximum Likelihood Estimation applied to Theta.}
\usage{
NMLE(rv, ips, Theta = 0, ccrit = 1e-04, max.iter = 2000)
}
\arguments{
\item{rv}{The response-vector for which Theta is to be estimated.}

\item{ips}{A table of item properties, where collumns represent the difficulty, discrimination, and guessing parameters (respectively).}

\item{ccrit}{The convergece criterion of Theta (i.e., desired degree of accuracy for the Theta estimate).}

\item{theta_hat}{the initial guestimate of Theta.}
}
\value{
An estimate for Theta which maximizes the likelihood of the observations.
}
\description{
An implementation of Newton's method for MLE of the person ability IRT parameter (aka. Theta). Finds the value of Theta which maximizes the probability of the observed response-pattern occuring, by substracting the derivative of the estimate from the estimate until successive improvements in the Theta estimates are inconsequentially small.
}
