% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{observedmoments}
\alias{observedmoments}
\title{Compute Moments of Observed Value Distribution.}
\usage{
observedmoments(x, type = c("raw", "central", "standardized"),
  orders = 4, correct = TRUE)
}
\arguments{
\item{x}{A vector of values, the distribution of which moments are to be calculated.}

\item{type}{A character vector determining which moment-types are to be calculated. Permissible values are "raw", "central", and "standardized".}

\item{orders}{The number of moment-orders to be calculated for each of the moment-types.}

\item{correct}{Whether to include bias correction in estimation of orders. Default is TRUE.}
}
\value{
A list of moment types, each a list of moment orders.
}
\description{
Computes Raw, Central, or Standardized moment properties of a vector of observed scores.
}
