% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTfunctions.R
\name{PCR}
\alias{PCR}
\title{Probability of Correct Response.}
\usage{
PCR(iptable, theta = 0)
}
\arguments{
\item{iptable}{A table of item properties, where column 1 represents the difficulty parameters, column 2 the discrimination parameters, and column 3 the guessing parameters of the items.}

\item{theta}{The value of Theta for which the probabilities for responding "1" is to be calculated.}
}
\value{
The probability for responding "1" on one or more items.
}
\description{
Calculates the probability of providing a response of "1" to a particular item with particular properties under 1, 2, and 3 parameter logistic IRT models, given a specific value of Theta.
}
