% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{rBetaMS}
\alias{rBetaMS}
\title{Random Draw from the Standard Beta PDD With Specific Mean and Variance.}
\usage{
rBetaMS(n, mean, variance = NULL, sd = NULL)
}
\arguments{
\item{n}{Number of observations to be drawn from under the Standard Beta PDD.}

\item{mean}{The mean of the target Standard Beta probability density distribution.}

\item{variance}{The variance of the target Standard Beta probability density distribution.}

\item{sd}{The standard deviation of the target Standard probability density distribution.}
}
\value{
A vector of length \code{n}, each value representing a random draw from the Standard Beta probability density distribution with defined mean and variance.
}
\description{
Draws random samples of observations from the Standard Beta probability density distribution with defined mean and variance.
}
