% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{rGammaBinom}
\alias{rGammaBinom}
\title{Random number generation under the generalized Binomial distribution (Gamma-Binomial) extended to non-integers.}
\usage{
rGammaBinom(n, size, prob)
}
\arguments{
\item{n}{Number of observations.}

\item{size}{Number of "trials" (zero or more). Need not be integer.}

\item{prob}{Probability of "success" on each "trial". Need not be integer.}
}
\description{
Random number generation under the generalized Binomial distribution (Gamma-Binomial) extended to non-integers.
}
\note{
Calls qGammaBinom(), which makes the random draw slower than what one might be used to (since qGammaBinom() calls pGammaBinom() and employs a search-algorithm to find the appropriate value).
}
\examples{
# Assume some variable follows a Gamma-Binomial distribution with
# "number of trials" = 10.5 and probability of "success" for each "trial"
# = 0.75 To draw a random value from this distribution:
rGammaBinom(n = 1, size = 10, prob = 0.75)
}
