% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{Beta.gfx.poly.cdf}
\alias{Beta.gfx.poly.cdf}
\title{Coordinate Generation for Marking an Area Under the Curve for the Standard Beta Cumulative Probability Density Distribution.}
\usage{
Beta.gfx.poly.cdf(from, to, by, alpha, beta)
}
\arguments{
\item{from}{The point of the x-axis from where to start producing y-density values.}

\item{to}{The point of the x-axis to where y-density values are to be produced.}

\item{by}{The resolution (or specing) at which to produce y-density values.}

\item{alpha}{The Alpha shape-parameter value for the Standard Beta probability density distribution.}

\item{beta}{The Beta shape-parameter fort he Standard Beta probability density distribution.}
}
\value{
A two-column matrix with density-values of y to plot against corresponding location values of x.
}
\description{
Plotting tool, producing a two-column matrix with values of \code{y} corresponding to locations on \code{x}. Useful for shading areas under the curve when tracing the line for the Standard Beta probability density function.
}
