% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{BetaMoments}
\alias{BetaMoments}
\title{Compute Moments of Beta Probability Density Distributions.}
\usage{
BetaMoments(A, B, TYPES = c("raw", "central", "standardized"),
  ORDERS = 4)
}
\arguments{
\item{A}{The Alpha shape parameter of the PDD.}

\item{B}{The Beta shape parameter of the PDD.}

\item{TYPES}{A character vector determining which moment-types are to be calculated. Permissible values are "raw", "central", and "standardized".}

\item{ORDERS}{The number of moment-orders to be calculated for each of the moment-types.}
}
\value{
A list of moment types, each a list of moment orders.
}
\description{
Computes Raw, Central, or Standardized moment properties of defined Standard Beta probability density distributions.
}
