% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{betamoments}
\alias{betamoments}
\title{Compute Moments of Beta Probability Density Distributions.}
\usage{
betamoments(a, b, mean = NULL, var = NULL, sd = NULL,
  types = c("raw", "central", "standardized"), orders = 4)
}
\arguments{
\item{a}{The Alpha shape parameter of the PDD.}

\item{b}{The Beta shape parameter of the PDD.}

\item{types}{A character vector determining which moment-types are to be calculated. Permissible values are "raw", "central", and "standardized".}

\item{orders}{The number of moment-orders to be calculated for each of the moment-types.}
}
\value{
A list of moment types, each a list of moment orders.
}
\description{
Computes Raw, Central, or Standardized moment properties of defined Standard Beta probability density distributions.
}
