% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{caStats}
\alias{caStats}
\title{Classification Accuracy Statistics.}
\usage{
caStats(tp, tn, fp, fn)
}
\arguments{
\item{tp}{The number or rate of true-positive classifications.}

\item{tn}{The number or rate of true-negative classifications.}

\item{fp}{The number or rate of false-positive classifications.}

\item{fn}{The number or rate of false-negative classifications.}
}
\value{
A list of classification accuracy statistics based on true/false positive/negative statistics. Specifically, the sensitivity, specificity, positive likelihood ratio, negative likelihood ratio, positive predictive value, negative predictive value, and Youden's J.
}
\description{
Provides a set of statistics often used for conveying information regarding the certainty of classifications based on tests.
}
