% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{dBeta.4P}
\alias{dBeta.4P}
\title{Probability Density under the Four-Parameter Beta PDD.}
\usage{
dBeta.4P(x, l, u, alpha, beta)
}
\arguments{
\item{x}{Value of X.}

\item{l}{The first (lower) location parameter.}

\item{u}{The second (upper) location paraeter.}

\item{alpha}{The first shape parameter.}

\item{beta}{The second shape parameter.}
}
\value{
The value for the probability density at specified values of X.
}
\description{
Gives the density at desired values of X under the Four-Parameter Beta PDD.
}
