% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{dBetaMS}
\alias{dBetaMS}
\title{Density under a specific point of the Standard Beta PDD with specific mean and variance.}
\usage{
dBetaMS(x, M, VAR = NULL, SD = NULL)
}
\arguments{
\item{x}{A specific point on the x-axis of the Standard Beta PDD.}

\item{M}{The mean of the target Standard Beta probability density distribution.}

\item{VAR}{The variance of the target Standard Beta probability density distribution.}

\item{SD}{The standard deviation of the target Standard Beta probability density distribution.}
}
\value{
A numeric value representing the required value for the Beta Shape-parameter in order to produce a Standard Beta probability density distribution with the target mean and variance.
}
\description{
Density under a specific point of the Standard Beta PDD with specific mean and variance.
}
