% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{LL.CA}
\alias{LL.CA}
\title{An implementation of the Livingston and Lewis (1995) Method to Classification Accuracy based on Test Scores.}
\usage{
LL.CA(x = NULL, min = 0, max = 1, reliability, cut, truecut = NULL)
}
\arguments{
\item{x}{A vector of observed scores for which a beta-distribution is to be fitted.}

\item{reliability}{The observed-score correlation with the true-score.}

\item{cut}{The cutoff value for classifying observations into pass or fail categories.}

\item{truecut}{Optional specification of a "true" cutoff. Useful for producing ROC curve values.}
}
\value{
A confusion matrix estimating the proportion of true/false pass/fail categorizations for a test, given a specific distribution of observed scores.
}
\description{
An implementation of what has been come to be known as the "Livingston and Lewis approach" to classification accuracy, which assumes that observed-scores, true-scores, and errors of measurement follow the four-parameter beta distribution. Under this assumption, the expected classification consistency and accuracy of tests can be estimated from observed outcomes and estimated test reliability.
}
\references{
Livinston, Samuel A. and Lewis, Charles. (1995). Estimating the Consistency and Accuracy of Classifications Based on Test Scores. Journal of Educational Measurement, 32(2).
}
