% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{LL.ROC}
\alias{LL.ROC}
\title{ROC curves for the Livingston and Lewis approach.}
\usage{
LL.ROC(x = NULL, min = 0, max = 1, reliability, truecut)
}
\arguments{
\item{x}{A vector of observed results.}

\item{min}{The minimum possible value to attain on the observed-score scale. Default is 0 (assuming \code{x} represent proportions).}

\item{max}{The maximum possible value to attain on the observed-score scale. Default is 1 (assuming \code{x} represent proportions).}

\item{reliability}{The reliability coefficient of the test.}

\item{truecut}{The true point along the x-scale that marks the categorization-threshold.}
}
\value{
A plot tracing the ROC curve for the test.
}
\description{
Generate a ROC curve plotting the false-positive rate against the true-positive rate at different cut-off values across the observed proportion-score scale.
}
