% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{MLB}
\alias{MLB}
\title{Most Likely True Beta Value Given Observed Outcome.}
\usage{
MLB(a, b, x = NULL, n = NULL)
}
\arguments{
\item{a}{Observed alpha value for fitted Standard Beta PDD.}

\item{b}{Observed beta value for fitted Standard Beta PDD.}

\item{x}{Observed proportion-correct outcome.}

\item{n}{Test-length.}
}
\value{
The Beta shape-parameter value for the Standard Beta probability density distribution where the observed mean is the expected mode.
}
\description{
Assuming a prior standard Beta distribution, return the Beta value where the observed mean becomes the mode.
}
