% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{MLM}
\alias{MLM}
\title{Most Likely Mean of the Standard Beta PDD, Given that the Observation is Considered the Most Likely Observation of the Standard Beta PDD (i.e., Mode).}
\usage{
MLM(a, b, x = NULL, n = NULL)
}
\arguments{
\item{a}{Observed alpha value for fitted Standard Beta PDD.}

\item{b}{Observed beta value for fitted Standard Beta PDD.}

\item{x}{Observed proportion-correct outcome.}

\item{n}{Test-length.}
}
\value{
The expected mean of the Standard Beta probability density distribution, for which the observed mean is the most likely value.
}
\description{
Assuming a prior standard Beta distribution, returns the expected mean of the distribution under the assumption that the observed value is the most likely value of the distribution.
}
