% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{MLM}
\alias{MLM}
\title{Most likely mean of the Standard Beta PDD, given that the observation is the most likely observation of the Standard Beta PDD (i.e., mode).}
\usage{
MLM(a, b)
}
\arguments{
\item{a}{Observed alpha value for fitted Standard Beta PDD.}

\item{b}{Observed beta value for fitted Standard Beta PDD.}
}
\value{
The expected mean of the Standard Beta probability density distribution, for which the observed mean is the most likely value.
}
\description{
Given a fitted Standard Beta Distribution, returns the expected mean of the distribution under the assumption that the observed value is the most likely value of the distribution.
}
