% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{qBeta.4P}
\alias{qBeta.4P}
\title{Quantile Given Probability Under the Four-Parameter Beta Distribution.}
\usage{
qBeta.4P(p, l, u, alpha, beta, lower.tail = TRUE)
}
\arguments{
\item{p}{A vector (or single value) of proportions or probabilities for which the corresponding value of \code{x} (i.e., the quantiles) are to be calculated.}

\item{l}{The first (lower) location parameter.}

\item{u}{The second (upper) location parameter.}

\item{alpha}{The first shape parameter.}

\item{beta}{The second shape parameter.}

\item{lower.tail}{Logical. Whether the quantile(s) to be calculated is to be under the lower or upper tail. Default is \code{TRUE} (lower tail).}
}
\value{
A vector of quantiles for specified probabilities or proportions of observations under the four-parameter Beta distribution.
}
\description{
Function for calculating the quantile (i.e., value of \code{x}) for a given proportion (i.e., the value of \code{y}) under the Four-Parameter Beta Distribution.
}
\examples{
# Assume some variable follows a four-parameter Beta distribution with
# location parameters l = 0.25 and u = 0.75, and shape parameters alpha = 5
# and beta = 3. To compute the quantile at a specific point of the
# distribution (e.g., 0.5) using qBeta.4P():
qBeta.4P(p = 0.5, l = 0.25, u = 0.75, alpha = 5, beta = 3)
}
