% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{rBetaMS}
\alias{rBetaMS}
\title{Random draw from the Standard Beta PDD with specific mean and variance.}
\usage{
rBetaMS(n, M, VAR = NULL, SD = NULL)
}
\arguments{
\item{n}{Number of observations to be drawn from under the Standard Beta PDD.}

\item{M}{The mean of the target Standard Beta probability density distribution.}

\item{VAR}{The variance of the target Standard Beta probability density distribution.}

\item{SD}{The standard deviation of the target Standard probability density distribution.}
}
\description{
Draws random samples of observations from the Standard Beta probability density distribution with defined mean and variance.
}
